+++ app/reviews/makeup/page.tsx
@@ -85,6 +85,12 @@ const Makeup = () => {
   const handleDelete = async (id: string) => {
     if (!isAuthenticated) return;
 
+    const confirmDelete = window.confirm("Are you sure you want to delete this post?");
+    
+    if (!confirmDelete) {
+      return; 
+    }
+
     try {
       const docRef = doc(db, 'makeup', id);
       await deleteDoc(docRef);
